ref dialogs

+++ main.py
@@ -127,7 +127,7 @@
 #
 # if __name__ == '__main__':
 #     main()
-
+import os
 from PyQt5.QtWidgets import *
 from PyQt5.QtGui import *
 from PyQt5.QtCore import *
@@ -146,7 +146,7 @@ SETTINGS_TRAY = 'settings/tray'
 class EditorApp(QMainWindow, ui.design.Ui_MainWindow):
 
     lines = None
-    filename = "Default" or None
+    filename = "Untitled*" or None
 
     def __init__(self):
         super().__init__()
@@ -164,12 +164,16 @@ class EditorApp(QMainWindow, ui.design.Ui_MainWindow):
 
         self.codeEditor = self.check_lines_state()
 
-        self.tabWidget.addTab(self.codeEditor, self.filename)
+        # self.tabWidget.addTab(self.codeEditor, self.filename)
+        self.tabWidget.tabCloseRequested.connect(lambda index: self.tabWidget.removeTab(index))
+        # self.tabWidget.insertTab(lambda index: self.tabWidget.addTab(index, self.filename))
         # self.verticalLayout.addWidget(self.codeEditor)
 
         self.action_Open.triggered.connect(self.file_open)
-        self.action_Save_File.triggered.connect(self.file_save)
+        self.action_Save.triggered.connect(self.save_file)
+        self.action_Save_as.triggered.connect(self.save_file)
         self.action_Exit.triggered.connect(qApp.quit)
+        self.action_New.triggered.connect(self.new_file)
 
     def check_lines_state(self):
         self.codeEditor = QCodeEditor(DISPLAY_LINE_NUMBERS=self.lines,
@@ -177,31 +181,70 @@ class EditorApp(QMainWindow, ui.design.Ui_MainWindow):
                                       SyntaxHighlighter=[x for x in (XMLHighlighter, PythonHighlighter)])
         return self.codeEditor
 
-    def file_save(self):
-        options = QFileDialog.Options()
-        # options |= QFileDialog.DontUseNativeDialog
-        file_name, _ = QFileDialog.getSaveFileName(self, "QFileDialog.getSaveFileName()", "",
-                                                   "All Files (*);;Python Files (*.py)", options=options)
-        if file_name:
-            with open(file_name, 'wb') as file:
+    def save_file(self):
+        if os.path.exists(self.filename):
+            with open(self.filename, 'wb') as file:
                 text = self.codeEditor.toPlainText()
                 file.write(text.encode())
-                file.close()
-            self.filename = file_name
-            self.tabWidget.addTab(self.codeEditor, self.filename)
+        else:
+            options = QFileDialog.Options()
+            file_name, _ = QFileDialog.getSaveFileName(self, "Save file as...", "",
+                                                       "Python Files (*.py)", options=options)
+            if file_name:
+                tab = QCodeEditor(DISPLAY_LINE_NUMBERS=self.lines,
+                                  HIGHLIGHT_CURRENT_LINE=True,
+                                  SyntaxHighlighter=[x for x in (XMLHighlighter, PythonHighlighter)])
+                with open(file_name, 'wb') as file:
+                    text = self.codeEditor.toPlainText()
+                    file.write(text.encode())
+                tab.setPlainText(text)
+                self.filename = file_name
+                tab_index = self.tabWidget.addTab(tab, self.filename)
+                try:
+                    self.tabWidget.removeTab(tab_index)
+                except Exception as ex:
+                    print(ex)
+                self.tabWidget.addTab(tab, self.filename)
+                tabs = self.tabWidget
+                self.codeEditor = tab
 
     def file_open(self):
         options = QFileDialog.Options()
-        # options |= QFileDialog.DontUseNativeDialog
-        file_name, _ = QFileDialog.getOpenFileName(self, "QFileDialog.getOpenFileName()", "",
+        file_name, _ = QFileDialog.getOpenFileName(self, "Open file", "",
                                                    "All Files (*);;Python Files (*.py)", options=options)
+
         if file_name:
+
+            tab = QCodeEditor(DISPLAY_LINE_NUMBERS=self.lines,
+                              HIGHLIGHT_CURRENT_LINE=True,
+                              SyntaxHighlighter=[x for x in (XMLHighlighter, PythonHighlighter)])
             with open(file_name, 'rb') as file:
                 text = file.read()
-                self.codeEditor.setPlainText(text.decode())
-            self.codeEditor.moveCursor(QTextCursor.End)
+                tab.setPlainText(text.decode())
+            tab.moveCursor(QTextCursor.End)
             self.filename = file_name
-            self.tabWidget.addTab(self.codeEditor, self.filename)
+
+            title = []
+
+            if file_name not in title:
+                global tab_index
+                self.tabWidget.addTab(tab, self.filename)
+                tab_index = self.tabWidget.addTab(tab, self.filename)
+                self.codeEditor = tab
+                self.tabWidget.setCurrentIndex(tab_index)
+                title.append(self.tabWidget.tabText(tab_index))
+            elif file_name in title:
+                for x in title:
+                    if x == file_name:
+                        self.tabWidget.setCurrentWidget(tab)
+
+    def new_file(self):
+        self.filename = "Untitled*" or None
+        tab = QCodeEditor(DISPLAY_LINE_NUMBERS=self.lines,
+                          HIGHLIGHT_CURRENT_LINE=True,
+                          SyntaxHighlighter=[x for x in (XMLHighlighter, PythonHighlighter)])
+        self.tabWidget.addTab(tab, self.filename)
+        self.codeEditor = tab
 
     # Слот для сохранения настроек чекбокса
     def save_check_box_settings(self):
